make the changes necessary to get usbloader to build with brainslug included
this will build, and run. I am able tolaunch Mario Kart fine
without any modules on the sd card.

if console-gx module is on the sd card,

Mario Kart will launch
The console will appear on the bottom with the usual
white WII REMOTE warning screen but upon pressing A to continue past
to the mario kart wii title screen, the upper part of the screen
goes black and the bottom console
shows existing all threads and a callback or return..

The wii remote is still connected,

pressing Up, 1, Down, 2, Right, -, Left, +, B, A
in that order does make the crash handler shows up.

special thanks to Leseratte for his help so far.